chore: type에서 role로 패키지 명 변경
