+++ src/main/java/chess/domain/piece/role/Bishop.java
@@ -1,4 +1,4 @@
-package chess.domain.piece.type;
+package chess.domain.piece.role;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
@@ -6,29 +6,27 @@ import chess.domain.piece.position.WayPoints;
 import chess.domain.piece.position.Path;
 import chess.domain.piece.position.PiecePosition;
 
-import java.util.Collections;
+public class Bishop extends Piece {
 
-public class King extends Piece {
-
-    public King(final Color color, final PiecePosition piecePosition) {
+    public Bishop(final Color color, final PiecePosition piecePosition) {
         super(color, piecePosition);
     }
 
     @Override
     protected void validateMovable(final Path path) {
-        if (!path.isUnitDistance()) {
-            throw new IllegalArgumentException("킹은 한칸만 이동할 수 있습니다.");
+        if (!path.isDiagonal()) {
+            throw new IllegalArgumentException("비숍은 대각선으로만 이동할 수 있습니다.");
         }
     }
 
     @Override
     protected WayPoints wayPointsWithCondition(final Path path) {
-        return new WayPoints(Collections.emptyList());
+        return new WayPoints(path.wayPoints());
     }
 
     @Override
     public boolean isKing() {
-        return true;
+        return false;
     }
 
     @Override
@@ -38,6 +36,6 @@ public class King extends Piece {
 
     @Override
     public double score() {
-        return 0;
+        return 3;
     }
 }
