+++ src/main/java/chess/domain/piece/role/King.java
@@ -1,28 +1,24 @@
-package chess.domain.piece.type;
+package chess.domain.piece.role;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
+import chess.domain.piece.position.WayPoints;
 import chess.domain.piece.position.Path;
 import chess.domain.piece.position.PiecePosition;
-import chess.domain.piece.position.WayPoints;
 
 import java.util.Collections;
 
-public class Knight extends Piece {
+public class King extends Piece {
 
-    public Knight(final Color color, final PiecePosition piecePosition) {
+    public King(final Color color, final PiecePosition piecePosition) {
         super(color, piecePosition);
     }
 
     @Override
     protected void validateMovable(final Path path) {
-        if (Math.abs(path.fileDistance()) == 1 && Math.abs(path.rankDistance()) == 2) {
-            return;
+        if (!path.isUnitDistance()) {
+            throw new IllegalArgumentException("킹은 한칸만 이동할 수 있습니다.");
         }
-        if (Math.abs(path.fileDistance()) == 2 && Math.abs(path.rankDistance()) == 1) {
-            return;
-        }
-        throw new IllegalArgumentException("나이트는 그렇게 움직일 수 없습니다.");
     }
 
     @Override
@@ -32,7 +28,7 @@ public class Knight extends Piece {
 
     @Override
     public boolean isKing() {
-        return false;
+        return true;
     }
 
     @Override
@@ -42,6 +38,6 @@ public class Knight extends Piece {
 
     @Override
     public double score() {
-        return 2.5;
+        return 0;
     }
 }
